Continue refactoring packages
