+++ common/conf/financial_transactions/queries_read.yml
