+++ common/conf/social_network/social_network.gql
