+++ common/configuration/BUILD
@@ -1,6 +1,13 @@
 java_library(
     name = "benchmark-configuration",
-    srcs = glob(["*.java"]),
+#    srcs = glob(["*.java"]),
+    srcs = [
+        "BenchmarkConfigurationFile.java",
+        "BenchmarkConfiguration.java",
+        "BenchmarkArguments.java",
+        "ConfigurationException.java",
+        "QueriesConfigurationFile.java"
+    ],
     deps = [
         "//dependencies/maven/artifacts/com/fasterxml/jackson/core:jackson-core",
         "//dependencies/maven/artifacts/com/fasterxml/jackson/core:jackson-databind",
