+++ profiler/src/QueryProfiler.java
@@ -1,117 +1,151 @@
-/*
- *  GRAKN.AI - THE KNOWLEDGE GRAPH
- *  Copyright (C) 2018 GraknClient Labs Ltd
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
 package grakn.benchmark.profiler;
 
-import brave.Tracing;
+
+import brave.Span;
+import brave.Tracer;
+import grakn.benchmark.common.analysis.InsertQueryAnalyser;
 import grakn.benchmark.common.configuration.BenchmarkConfiguration;
-import grakn.benchmark.profiler.util.ProfilerException;
 import grakn.core.client.GraknClient;
-import grakn.core.concept.answer.Numeric;
+import grakn.core.concept.answer.Answer;
+import grakn.core.concept.answer.ConceptMap;
 import graql.lang.Graql;
+import graql.lang.query.GraqlCompute;
+import graql.lang.query.GraqlDelete;
+import graql.lang.query.GraqlGet;
+import graql.lang.query.GraqlInsert;
 import graql.lang.query.GraqlQuery;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.LinkedList;
 import java.util.List;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
+import java.util.Set;
 import java.util.stream.Collectors;
 
-import static graql.lang.Graql.var;
-
-
-/**
- *
- */
-public class QueryProfiler {
-
+class QueryProfiler implements Runnable {
     private static final Logger LOG = LoggerFactory.getLogger(QueryProfiler.class);
 
-    private final BenchmarkConfiguration config;
-
+    private int concurrentId;
+    private String configName;
+    private String description;
+    private String dataGenerator;
+    private Tracer tracer;
     private final List<GraqlQuery> queries;
-    private final GraknClient client;
-    private final List<String> keyspaces;
-    private final int concurrentClients;
-    private ExecutorService executorService;
+    private final int repetitions;
+    private final int numConcepts;
+    private final GraknClient.Session session;
+    private final boolean deleteInsertedConcepts;
+    private final boolean traceDeleteInsertedConcepts;
+    private String executionName;
 
-    public QueryProfiler(GraknClient client, List<String> keyspaces, BenchmarkConfiguration config) {
-        this.config = config;
-        this.client = client;
-        this.keyspaces = keyspaces;
-        this.concurrentClients = config.concurrentClients();
-
-
-        // convert Graql strings into GraqlQuery types
-        this.queries = config.getQueries().stream()
-                .map(q -> (GraqlQuery) Graql.parse(q))
-                .collect(Collectors.toList());
-
-
-        // create 1 thread per client session
-        executorService = Executors.newFixedThreadPool(concurrentClients);
+    public QueryProfiler(BenchmarkConfiguration config, int concurrentId, Tracer tracer, List<GraqlQuery> queries, int repetitions, int numConcepts, GraknClient.Session session) {
+        configName = config.configName();
+        description = config.configDescription();
+        executionName = config.executionName();
+        deleteInsertedConcepts = config.deleteInsertedConcepts();
+        traceDeleteInsertedConcepts = config.traceDeleteInsertedConcepts();
+        dataGenerator = config.dataGenerator();
+        this.concurrentId = concurrentId;
+        this.tracer = tracer;
+        this.queries = queries;
+        this.repetitions = repetitions;
+        this.numConcepts = numConcepts;
+        this.session = session;
     }
 
-    public void processStaticQueries(int numRepeats, int numConcepts) {
-        LOG.trace("Starting processStaticQueries");
-        this.processQueries(queries, numRepeats, numConcepts);
-        LOG.trace("Finished processStaticQueries");
-    }
-
-    public int aggregateCount(GraknClient.Session session) {
-        try (GraknClient.Transaction tx = session.transaction().read()) {
-            List<Numeric> count = tx.execute(Graql.match(var("x").isa("thing")).get().count());
-            return count.get(0).number().intValue();
-        }
-    }
-
-    void processQueries(List<GraqlQuery> queries, int repetitions, int numConcepts) {
-        List<Future> runningConcurrentQueries = new LinkedList<>();
-
-        long start = System.currentTimeMillis();
-
-        for (int i = 0; i < concurrentClients; i++) {
-            // TODO: this can probably be optimised (keeping sessions open)
-            String keyspace = (keyspaces.size() > 1) ? keyspaces.get(i) : keyspaces.get(0);
-            GraknClient.Session session = client.session(keyspace);
-            ConcurrentQueries processor = new ConcurrentQueries(config, i, Tracing.currentTracer(), queries, repetitions, numConcepts, session);
-            runningConcurrentQueries.add(executorService.submit(processor));
-        }
-
-        // wait until all threads have finished
+    @Override
+    public void run() {
         try {
-            for (Future future : runningConcurrentQueries) {
-                future.get();
+            Span concurrentExecutionSpan = tracer.newTrace().name("concurrent-execution");
+            concurrentExecutionSpan.tag("configurationName", configName);
+            concurrentExecutionSpan.tag("description", description);
+            concurrentExecutionSpan.tag("executionName", executionName);
+            concurrentExecutionSpan.tag("concurrentClient", Integer.toString(concurrentId));
+            concurrentExecutionSpan.tag("graphGeneratorDefinition", dataGenerator);
+            concurrentExecutionSpan.tag("queryRepetitions", Integer.toString(repetitions));
+            concurrentExecutionSpan.tag("graphScale", Integer.toString(numConcepts));
+            concurrentExecutionSpan.start();
+
+            int counter = 0;
+            long startTime = System.currentTimeMillis();
+            for (int rep = 0; rep < repetitions; rep++) {
+                for (GraqlQuery query : queries) {
+
+                    if (counter % 100 == 0) {
+                        System.out.println("Executed query #: " + counter + ", elapsed time " + (System.currentTimeMillis() - startTime));
+                    }
+                    Span querySpan = tracer.newChild(concurrentExecutionSpan.context());
+
+                    if (query instanceof GraqlInsert) { querySpan.name("insert-query"); }
+                    else if (query instanceof GraqlGet) { querySpan.name("get-query"); }
+                    else if (query instanceof GraqlDelete) { querySpan.name("delete-query"); }
+                    else if (query instanceof GraqlCompute) { querySpan.name("compute-query"); }
+                    else { querySpan.name("query"); }
+
+                    querySpan.tag("query", query.toString());
+                    querySpan.tag("repetitions", Integer.toString(repetitions));
+                    querySpan.tag("repetition", Integer.toString(rep));
+                    querySpan.start();
+
+                    // perform trace in thread-local storage on the client
+                    Set<String> insertedConceptIds = null;
+                    try (Tracer.SpanInScope span = tracer.withSpanInScope(querySpan)) {
+                        // open new transaction
+                        GraknClient.Transaction tx = session.transaction().write();
+                        List<? extends Answer> answer = tx.execute(query);
+
+                        if (query instanceof GraqlInsert) {
+                            insertedConceptIds = InsertQueryAnalyser.getInsertedConcepts((GraqlInsert)query, (List<ConceptMap>)answer)
+                                        .stream().map(concept -> concept.id().toString())
+                                        .collect(Collectors.toSet());
+                        }
+                        tx.commit();
+                    } catch (Exception e) {
+                        LOG.error(
+                                "Exception in a concurrent query executor, query: " + query +
+                                ". Ensure every query is valid and each inserted concept is associated with an explicit variable",
+                                e);
+                        throw e;
+                    } finally {
+                        querySpan.finish();
+                    }
+
+
+                    if (deleteInsertedConcepts && insertedConceptIds != null) {
+                        if (traceDeleteInsertedConcepts) {
+                            Span deleteQuerySpan = tracer.newChild(concurrentExecutionSpan.context());
+                            deleteQuerySpan.name("delete-query");
+                            deleteQuerySpan.start();
+                            try (Tracer.SpanInScope span = tracer.withSpanInScope(deleteQuerySpan)) {
+                                // create one tx per ID in case the concept no longer exists and throws an error (attr dedup)
+                                GraknClient.Transaction tx = session.transaction().write();
+                                for (String conceptId : insertedConceptIds) {
+                                    tx.execute(Graql.parse("match $x id " + conceptId + "; delete $x;").asDelete());
+                                }
+                                tx.commit();
+                            } finally {
+                                deleteQuerySpan.finish();
+                            }
+                        } else {
+                            GraknClient.Transaction tx = session.transaction().write();
+                            insertedConceptIds.iterator().forEachRemaining(id -> tx.execute(Graql.parse("match $x id " + id + "; delete $x;").asDelete()));
+                            tx.commit();
+                        }
+                    }
+                    counter++;
+                }
             }
-        } catch (InterruptedException | ExecutionException e) {
-            throw new ProfilerException("Error in execution of profiled queries", e);
+
+            concurrentExecutionSpan.finish();
+            // give zipkin reporter time to finish transmitting spans/close spans cleanly
+            Thread.sleep(1000);
+        } catch (InterruptedException e) {
+            System.out.println("Thread sleeps during data generation were interrupted");
+            e.printStackTrace();
+            Thread.currentThread().interrupt();
+        } finally {
+            session.close();
         }
-
-        long length = System.currentTimeMillis() - start;
-        System.out.println("Time: " + length);
-    }
-
-    public void cleanup() {
-        executorService.shutdown();
+        System.out.println("Thread runnable finished running queries");
+        System.out.print("\n\n");
     }
 }
-
